Prevent Clang warning, second try

+++ src/xerus/measurments.cpp
@@ -493,7 +493,7 @@ namespace xerus {
                 } else { init = false; }
                 
                 // Rebuild stack
-                for(long i = degree()-1-long(unchangedModes); i >= 0; --i) {
+                for(long i = long(degree())-1-long(unchangedModes); i >= 0; --i) {
 					stack[size_t(i)](k&0) = stack[size_t(i+1)](k&1, l) * positions[j][size_t(i)](l);
 					stack[size_t(i)].reduce_representation();
                 }
